% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightHIPPO.R
\name{cut_hierarchy}
\alias{cut_hierarchy}
\title{Clustering pruning}
\usage{
cut_hierarchy(cluster_res, K)
}
\arguments{
\item{cluster_res}{output from lightHIPPO}

\item{K}{maximum number of clusters}
}
\value{
A vector of cluster labels.
}
\description{
This function will take the HIPPO result as input and return clustering labels with the desired number of clusters.
}
