% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightHIPPO.R
\name{lightHIPPO}
\alias{lightHIPPO}
\title{lightHIPPO}
\usage{
lightHIPPO(
  dat,
  K = 10,
  initial.round = 0,
  stop_at = 500,
  correctByK = FALSE,
  override.Zscore.cutoff = NULL,
  smallest.cluster.num = 200,
  random.num = 2500
)
}
\arguments{
\item{dat}{input data matrix (dense or sparse)}

\item{K}{maximum number of clusters}

\item{initial.round}{HIPPO rounds using a subset of features, default is 0}

\item{stop_at}{when initial.round > 0, each round will select up to this number of features for clustering and it will not go over all the features}

\item{correctByK}{whether taking into account the number of clusters when calculating the z-score cut-off, default is FALSE}

\item{override.Zscore.cutoff}{a pre-specified cut-off on Z-scores, the default is NULL.}

\item{smallest.cluster.num}{smallest number of cells required in the cluster. Clusters with cells smaller than this number will not be selected for further clustering.}

\item{random.num}{if smaller than the number of features, only this random set of features will be used to track numbers of inflated genes in each cluster, which will be used to determine the next cluster for further breakdown.}
}
\value{
A list with clustering results, selected features and z-scores.
\itemize{
  \item next_round_IDs - the current cluster labels
  \item sequence - the sequence of hierarchical clusters
  \item selected.gene.list - a list of IDs of selected genes  at each round
  \item selected.gene.Zscore - a list of z-scores of selected genes at each round
}
}
\description{
lightHIPPO
}
