% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightHIPPO.R
\name{organizing_hippo_features}
\alias{organizing_hippo_features}
\title{Post-process HIPPO Features}
\usage{
organizing_hippo_features(hippo.res)
}
\arguments{
\item{hippo.res}{output from lightHIPPO}
}
\value{
A list of selected features and z-scores that are ranked by their significance.
\itemize{
  \item ID - a list of IDs of selected genes at each round
  \item Zscore - a list of z-scores of selected genes at each round
}
}
\description{
This function will take the HIPPO result as input and return selected features for each round.
It will first identify genes that are selected at each round and define those as common features. These genes remain heterogeneous at each round, usually are not interesting. We remove them from the final curated gene list.
The function will then identify genes that are private to each round, defined as genes that are only inflated at round $k$, but no longer inflated in later rounds. These genes carry information for separation at round $k$, but their heterogeneity got reconciled by the newly introduced cluster.
}
